refactor: class & package name refactoring
클래스와 패키지 이름을 좀 더 적절하게 변경했다.
