--- src/main/java/baseball/inputchecking/InputChecking.java
@@ -1,6 +1,0 @@
-package baseball.inputchecking;
-
-public interface InputChecking {
-
-	public void inputCheck(String input);
-}
