+++ src/main/java/baseball/validator/InputValidator.java
@@ -1,0 +1,6 @@
+package baseball.validator;
+
+public interface InputValidator {
+
+	public void inputCheck(String input);
+}
